refactor(lite): extract streaming classification and registry into streaming.ts
